Day 10/agent_backend/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
httpx
python-dotenv
beautifulsoup4
langchain
langchain-community
google-generativeai
motor